@@ -4,13 +4,13 @@ version: 0.6.3
 homepage: https://github.com/ikicodedev/calendar_timeline
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  scrollable_positioned_list: ^0.1.7
-  intl: ^0.16.1
+  scrollable_positioned_list: ^0.1.10
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
