@@ -35,7 +35,6 @@ class HomePage extends StatefulWidget {
 }
 
 class _HomePageState extends State<HomePage> {
-
   DateTime _selectedDate;
 
   @override
@@ -71,9 +70,7 @@ class _HomePageState extends State<HomePage> {
               firstDate: DateTime.now(),
               lastDate: DateTime.now().add(Duration(days: 365)),
               onDateSelected: (date) {
-                setState(() {
-                  _selectedDate = date;
-                });
+                setState(() => _selectedDate = date);
               },
               leftMargin: 20,
               monthColor: Colors.white70,
@@ -83,19 +80,33 @@ class _HomePageState extends State<HomePage> {
               activeBackgroundDayColor: Colors.redAccent[100],
               dotsColor: Color(0xFF333A47),
               selectableDayPredicate: (date) => date.day != 23,
-              locale: 'en',
+              // locale: 'en',
             ),
             SizedBox(height: 20),
             Padding(
               padding: const EdgeInsets.only(left: 16),
-              child: FlatButton(
-                color: Colors.teal[200],
-                child: Text('RESET', style: TextStyle(color: Color(0xFF333A47))),
+              child: TextButton(
+                style: ButtonStyle(
+                  foregroundColor: MaterialStateProperty.all<Color>(
+                    Colors.teal[200],
+                  ),
+                ),
+                child: Text(
+                  'RESET',
+                  style: TextStyle(
+                    color: Color(0xFF333A47),
+                  ),
+                ),
                 onPressed: () => setState(() => _resetSelectedDate()),
               ),
             ),
             SizedBox(height: 20),
-            Center(child: Text('Selected date is $_selectedDate', style: TextStyle(color: Colors.white)))
+            Center(
+              child: Text(
+                'Selected date is $_selectedDate',
+                style: TextStyle(color: Colors.white),
+              ),
+            ),
           ],
         ),
       ),
